+++ Commands/PrintQueuedRequests.php
@@ -31,8 +31,11 @@ class PrintQueuedRequests extends ConsoleCommand
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $systemCheck = new SystemCheck();
-        $systemCheck->checkRedisIsInstalled();
+        $settings = Queue\Factory::getSettings();
+        if ($settings->isRedisBackend()) {
+            $systemCheck = new SystemCheck();
+            $systemCheck->checkRedisIsInstalled();
+        }
 
         $backend = Queue\Factory::makeBackend();
         $manager = Queue\Factory::makeQueueManager($backend);