+++ tests/Integration/Queue/ProcessorTest.php
@@ -32,6 +32,7 @@ class TestProcessor extends Processor {
  */
 class ProcessorTest extends IntegrationTestCase
 {
+    protected $testRequiresRedis = false;
 
     /**
      * @var TestProcessor
@@ -44,9 +45,9 @@ class ProcessorTest extends IntegrationTestCase
     private $queue;
 
     /**
-     * @var Redis
+     * @var Queue\Backend
      */
-    private $redis;
+    private $backend;
 
     /**
      * @var Queue\Lock
@@ -57,11 +58,11 @@ class ProcessorTest extends IntegrationTestCase
     {
         parent::setUp();
 
-        $this->redis = $this->createRedisBackend();
+        $this->backend = $this->createMySQLBackend();
 
-        $this->lock = new Queue\Lock($this->redis);
+        $this->lock = new Queue\Lock($this->backend);
 
-        $this->queue = new Queue\Manager($this->redis, $this->lock);
+        $this->queue = new Queue\Manager($this->backend, $this->lock);
         $this->queue->setNumberOfAvailableQueues(1);
         $this->queue->setNumberOfRequestsToProcessAtSameTime(3);
 
@@ -70,7 +71,7 @@ class ProcessorTest extends IntegrationTestCase
 
     public function tearDown()
     {
-        $this->clearRedisDb();
+        $this->clearBackend();
         parent::tearDown();
     }