+++ tests/Integration/Tracker/HandlerTest.php
@@ -78,7 +78,7 @@ class HandlerTest extends IntegrationTestCase
 
     public function tearDown()
     {
-        $this->clearRedisDb();
+        $this->clearBackend();
         parent::tearDown();
     }